% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgLambda.R
\name{avgLambda}
\alias{avgLambda}
\title{Convert Wavelengths to Integer Values and Average Corresponding Absorbances}
\usage{
avgLambda()
}
\value{
The original files are overwritten with the modified files.
}
\description{
This function reads a csv file containing columns of wavelength and
absorbances.  It rounds the wavelengths to integers and replaces the
absorbances corresponding to the rounded values with their averages.  NOTE
THAT ALL THE csv FILES IN THE CURRENT DIRECTORY ARE PROCESSED AND REPLACED
WITH THE MODIFIED FILEs.  You should use this function on a copy of the
directory.
}
\seealso{
\code{\link{gatherSpecFiles}} which is the function the user should
call.
}
\author{
Bryan A. Hanson, DePauw University
}
\keyword{utilities}
