% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepCIEgradient.R
\name{prepCIEgradient}
\alias{prepCIEgradient}
\title{Compute a gradient to fill the CIE chromaticity diagram}
\usage{
prepCIEgradient(vertices = NULL, colSpace = "sRGB", ex = 1, ...)
}
\arguments{
\item{vertices}{The vertices of a polygon that is to be filled with
the gradient.}

\item{colSpace}{Character.  The color space model to use.}

\item{ex}{Numeric.  The exposure factor. This shifts the gradient.
Be extremely careful with this.  See \code{\link{plotCIEchrom}}
for full details.}

\item{\ldots}{Arguments to be passed downstream.}
}
\value{
An array containing the data needed to draw the gradient.
}
\description{
This function creates a gradient to fill the CIE chromaticity diagram.
}
\author{
Bryan A. Hanson, DePauw University. \email{hanson@depauw.edu}
}
\seealso{
\code{\link{plotCIEchrom}} for examples of this function in use.
}
\keyword{utilities}

