% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SpecHelpers-package.R
\docType{package}
\name{SpecHelpers-package}
\alias{SpecHelpers}
\alias{SpecHelpers-package}
\title{SpecHelpers: Spectroscopy related utilities}
\description{
Utility functions for spectroscopy.  1. Functions to simulate spectra for
use in teaching or testing.  2.  Functions to process files created by
LoggerPro and SpectraSuite software.
}
\author{
Bryan A. Hanson, DePauw University, Greencastle Indiana USA

Maintainer: Bryan A. Hanson \email{hanson@depauw.edu}
}
\keyword{package}

