% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecHelpers-package.R
\docType{package}
\name{SpecHelpers-package}
\alias{SpecHelpers-package}
\alias{SpecHelpers}
\title{Spectroscopy Related Utilities
Maintainer: Bryan A. Hanson \email{hanson@depauw.edu}}
\description{
SpecHelpers: Spectroscopy Related Utilities
}
\details{
Utility functions for spectroscopy.  1. Functions to simulate spectra for
use in teaching or testing.  2.  Functions to process files created by
LoggerPro and SpectraSuite software.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bryanhanson/SpecHelpers}
  \item Report bugs at \url{https://github.com/bryanhanson/SpecHelpers/issues}
}

}
\author{
Bryan A. Hanson, DePauw University, Greencastle Indiana USA
}
\keyword{package}
